Sales Dataset
The dataset "Sales Analysis Report" contains sales order information with various details for each order. Here's a breakdown of the columns present in the dataset:

OrderNumber: A unique identifier for each sales order.
OrderDate: The date when the order was placed.
Ship Date: The date when the order was shipped.
Customer Name Index: An index or identifier representing the customer who placed the order.
Channel: The sales channel through which the order was placed (e.g., Wholesale, Distributor, Export).
Currency Code: The currency code for the transaction (e.g., NZD, USD).
Warehouse Code: A code identifying the warehouse from which the goods were shipped.
Delivery Region Index: An index or identifier representing the region to which the order was delivered.
Product Description Index: An index or identifier representing the product ordered.
Order Quantity: The quantity of the product ordered.
Unit Selling Price: The selling price per unit of the product.
Unit Cost: The cost per unit of the product.
The dataset seems to track the sales orders, including details about the orders, customers, products, and logistics like shipping and delivery. The information is useful for analyzing sales performance, customer behavior, product popularity, and logistical efficiency across different channels and regions. ​